The Note element now vertically stack controls, similar to a StackPanel.
Added padding/positioning attribute support to the Note element.

Fixed the Refresh button on iOS to remain at the top of the screen and not
be overlapped by content.

Fixed an issue causing Paragraph and Quote text to take on the exact formatting of the XML (rather than a logical layout)

Fixed an issue causing the keyboard to appear after change orientations on Android devices.