Added support for MarginRight and MarginBottom
Cleaned up and simplified Padding and Margin parsing for controls.

Added support for Margins to all controls.

Margins are now a style rather than an attribute.

Fixed an issue causing Header Top margins to expand the height of the container.

When container objects don't have ChildAlignment set, default to left, not the container's alignment.
-Fixes an issue causing children in containers to be aligned incorrectly.

Fixed an issue causing the Android Share Panel to open for as many times as the Share button had been pressed since the app loaded.

Fixed an issue causing Android Springboard buttons to activate when the springboard was closed.

Sample_Note.xml will now be loaded when running the app in Debug mode.

Fixed an issue on iOS 8 causing saved note states to be lost.
Fixed an issue on iOS 8 causing the orientation change animation to be skipped.