Added a custom date picker for Android, which fixes an issue causing the calendar version of the date picker to always show on Android 5.0+.
Updated quotes to require the citation to be tapped to view a verse.

Citations that have a URL associated are now underlined.

Fixed an issue causing a tint on the iOS Navtoolbar to cause the coloring to be inaccurate.

Fixed many inconsistent colors on Android.

Added source fonts and art.