Updated GroupFinder to display more useful information. (Distance, Meeting Time)
GroupFinder now colors the closest group.

Improved GroupFinder navigation when tapping an annotation or list item.

Updated Android ControlStyling to allow styling layers that don't have both top and bottom borders.

Fixed the Android Back Button to work more intuitively for the user.

Rewrote Android Springboard Panning to be intuitive and less buggy.

Fixed layout issues in the Android Login screen.

Updated Prayer Creation to color required fields red if Submit was pressed with them empty.

Don't show the NavToolbar on the GroupFinder page. It's annoying.

Fixed an issue causing the ImageCropFragment to be layed out incorrectly if launched in landscape orientation.

Fixed crashing related to spamming Facebook / Cancel on the Login Page on BOTH platforms. (Yeesh)

Fixed iOS Back Button issues that allowed the back button when in landscape mode. It's FINALLY working well.

Fixed MultiTouch issues on iOS that would allow changing activities when the springboard was closed, which was weird.